requirements.txt
aiogram
SQLAlchemy==2.0.15
APScheduler==3.10.1
PyYAML==6.0
openai==0.27.0
pandas==2.1.0
openpyxl==3.1.2
numpy==1.24.3
psycopg2-binary
python-docx
jinja2